+++ src/Giraffe/XmlViewEngine.fs
@@ -1,4 +1,3 @@
-
 /// ---------------------------
 /// Attribution to original authors of this code
 /// ---------------------------