+++ tests/Giraffe.Tests/HttpHandlerTests.fs
@@ -11,13 +11,9 @@ open Microsoft.Extensions.Primitives
 open Microsoft.Extensions.Logging
 open Xunit
 open NSubstitute
-open Giraffe.HttpHandlers
-open Giraffe.Middleware
-open Giraffe.XmlViewEngine
-open Giraffe.DotLiquid.HttpHandlers
-open Giraffe.Tests.Asserts
-open Giraffe.Tasks
 open Newtonsoft.Json
+open Giraffe.XmlViewEngine
+open Giraffe.Tests.Asserts
 
 // ---------------------------------
 // Helper functions
@@ -273,41 +269,6 @@ let ``PUT "/post/2" returns 404 "Not found"`` () =
             let body = getBody ctx
             Assert.Equal(expected, body)
             Assert.Equal(404, ctx.Response.StatusCode)
-    }
-
-[<Fact>]
-let ``GET "/dotLiquid" returns rendered html view`` () =
-    let ctx = Substitute.For<HttpContext>()
-    let dotLiquidTemplate =
-        "<html><head><title>DotLiquid</title></head>" +
-        "<body><p>{{ foo }} {{ bar }} is {{ age }} years old.</p>" +
-        "</body></html>"
-
-    let obj = { Foo = "John"; Bar = "Doe"; Age = 30 }
-
-    let app =
-        choose [
-            GET >=> choose [
-                route "/"          >=> text "Hello World"
-                route "/dotLiquid" >=> dotLiquid "text/html" dotLiquidTemplate obj ]
-            POST >=> choose [
-                route "/post/1"    >=> text "1" ]
-            setStatusCode 404      >=> text "Not found" ]
-
-    ctx.Request.Method.ReturnsForAnyArgs "GET" |> ignore
-    ctx.Request.Path.ReturnsForAnyArgs (PathString("/dotLiquid")) |> ignore
-    ctx.Response.Body <- new MemoryStream()
-    let expected = "<html><head><title>DotLiquid</title></head><body><p>John Doe is 30 years old.</p></body></html>"
-
-    task {
-        let! result = app next ctx
-
-        match result with
-        | None -> assertFailf "Result was expected to be %s" expected
-        | Some ctx ->
-            let body = getBody ctx
-            Assert.Equal(expected, body)
-            Assert.Equal("text/html", ctx.Response |> getContentType)
     }
 
 [<Fact>]