+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@ Release Notes
 
 - Added HTTP status code helper functions.
 - Added `defaultSerializeJson` and `defaultDeserializeJson` methods.
+- Auto opened default Giraffe modules so that the core functionality can be entirely consumed through a single `open Giraffe` statement.
+- The functionality from `Giraffe.Razor.Middleware` and `Giraffe.Razor.HttpHandlers` can be both consumed through a single `open Giraffe.Razor` now.
 
 ### Bug fixes
 
@@ -26,6 +28,7 @@ Release Notes
     - `ctx.WriteText` --> `ctx.WriteTextAsync`
     - `ctx.RenderHtml` --> `ctx.RenderHtmlAsync`
     - `ctx.ReturnHtmlFile` --> `ctx.ReturnHtmlFileAsync`
+- Renamed `Giraffe.DotLiquid.HttpHandlers` module to `Giraffe.DotLiquid`
 
 # 0.1.0-beta-310